Align more closely to Backbone API
A bit counterintuitive, but without these changes the Pegasus'ed Backbone.sync lacks its documented return value, and the arguments of options.success and options.error and the according events differ from the un-pegasus'ed variants. IMHO the Backbone documentation is a bit misleading about this, but just compare the arguments with and without Pegasus to verify that this works. This PR relates to #1, actually we're colleagues.